Refactor controller and service objects
