+++ spec/services/order_builder_spec.rb
@@ -1,16 +1,17 @@
 require 'rails_helper'
 
-RSpec.describe PurchasesHandler::CartFinder, type: :request do
+RSpec.describe PurchasesHandler::OrderBuilder, type: :request do
   include Requests
 
   describe "#call" do
     subject(:request!) { post '/purchases', params: params }
 
-    context "when cart exists" do
+    context "when order is created", skip: true do
       let(:gateway) { :paypal }
-      let(:user) { create(:user) }
       let(:params) { { gateway: gateway, cart_id: cart_id } }
-      let!(:cart_id) { create(:cart, user: user).id }
+      let(:user) { create(:user) }
+      let(:cart) { create(:cart, user: user) }
+      let(:cart_id) { cart.id }
 
       before { request! }
 
@@ -18,6 +19,7 @@ RSpec.describe PurchasesHandler::CartFinder, type: :request do
         expect(response).to have_http_status(:ok)  
       end
     end
+
   end
   
 end
